% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_classes.R
\name{as_tibble.bid_stage}
\alias{as_tibble.bid_stage}
\title{Convert bid_stage to tibble}
\usage{
\method{as_tibble}{bid_stage}(x, ...)
}
\arguments{
\item{x}{A bid_stage object}

\item{...}{Additional arguments (unused)}
}
\value{
A tibble
}
\description{
Convert bid_stage to tibble
}
