% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bid_concepts.R
\name{bid_concepts}
\alias{bid_concepts}
\title{List all BID Framework Concepts}
\usage{
bid_concepts(search = NULL)
}
\arguments{
\item{search}{Optional search term to filter concepts. Multiple terms can be
provided separated by commas or spaces.}
}
\value{
A tibble with concept details.
}
\description{
This function returns a tibble listing all the key concepts in the Behavior
Insight Design framework, including their descriptions, categories,
references, and example applications.
}
\examples{
bid_concepts()
bid_concepts("cognitive")
bid_concepts("visual, hierarchy") # Multiple search terms

}
