% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bid_get_quiet}
\alias{bid_get_quiet}
\title{Get current quiet mode setting}
\usage{
bid_get_quiet()
}
\value{
Logical indicating whether quiet mode is enabled
}
\description{
Check whether bidux is currently in quiet mode.
}
\examples{
# Check current quiet setting
bid_get_quiet()

}
