% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bid-classes.R
\name{bid_result}
\alias{bid_result}
\title{Constructor for BID result collection objects}
\usage{
bid_result(stages)
}
\arguments{
\item{stages}{List of bid_stage objects}
}
\value{
Object of class 'bid_result'
}
\description{
Constructor for BID result collection objects
}
