% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_messaging.R
\name{bid_set_quiet}
\alias{bid_set_quiet}
\title{Set global quiet mode for bidux functions}
\usage{
bid_set_quiet(quiet = TRUE)
}
\arguments{
\item{quiet}{Logical indicating whether to enable quiet mode.
When TRUE, most bidux messages are suppressed.}
}
\value{
The previous value of the quiet option (invisibly)
}
\description{
Convenience function to set the global quiet option for all bidux functions.
When quiet mode is enabled, most informational messages are suppressed.
}
\examples{
# Enable quiet mode
bid_set_quiet(TRUE)

# Disable quiet mode
bid_set_quiet(FALSE)

}
