% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_classes.R
\name{bid_stage}
\alias{bid_stage}
\title{Constructor for BID stage objects}
\usage{
bid_stage(stage, data, metadata = list())
}
\arguments{
\item{stage}{Character string indicating the stage name}

\item{data}{Tibble containing the stage data}

\item{metadata}{List containing additional metadata}
}
\value{
Object of class 'bid_stage'
}
\description{
Constructor for BID stage objects
}
