% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stage.R
\name{classify_audience}
\alias{classify_audience}
\title{classify audience type from description text}
\usage{
classify_audience(audience_text)
}
\arguments{
\item{audience_text}{character string describing the audience}
}
\value{
character string: one of "executive", "analyst", "marketing", "operations", "general"
}
\description{
classify audience type from description text
}
\keyword{internal}
