% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_bid_result}
\alias{create_bid_result}
\title{Create consistent result structure (DRY principle)}
\usage{
create_bid_result(
  data_list,
  class_name,
  attributes = list(),
  return_tibble = TRUE
)
}
\arguments{
\item{data_list}{List of data elements}

\item{class_name}{S3 class name to apply}

\item{attributes}{Named list of attributes to set}

\item{return_tibble}{Whether to return tibble if available}
}
\description{
Standardized result creation to reduce duplication
}
\keyword{internal}
