% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telemetry_notices.R
\name{create_telemetry_notice}
\alias{create_telemetry_notice}
\title{Create a telemetry notice from problem and evidence}
\usage{
create_telemetry_notice(central_question, problem, evidence)
}
\arguments{
\item{central_question}{The central question for the interpret stage}

\item{problem}{Problem description string}

\item{evidence}{Evidence description string}
}
\value{
bid_stage object in the Notice stage
}
\description{
Create a telemetry notice from problem and evidence
}
\keyword{internal}
