% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telemetry_notices.R
\name{.create_issues_tibble}
\alias{.create_issues_tibble}
\title{Create tidy issues tibble from notice issues list}
\usage{
.create_issues_tibble(notice_issues, total_sessions, events)
}
\arguments{
\item{notice_issues}{List of bid_stage objects from telemetry analysis}

\item{total_sessions}{Total number of sessions analyzed}

\item{events}{Raw events data frame}
}
\value{
Tibble with structured issue metadata
}
\description{
Create tidy issues tibble from notice issues list
}
\keyword{internal}
