% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_classes.R
\name{extract_stage}
\alias{extract_stage}
\title{Extract specific stage from bid_result}
\usage{
extract_stage(workflow, stage)
}
\arguments{
\item{workflow}{A bid_result object}

\item{stage}{Character string with stage name}
}
\value{
A bid_stage object or NULL if not found
}
\description{
Extract specific stage from bid_result
}
