% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telemetry.R
\name{find_unused_inputs}
\alias{find_unused_inputs}
\title{Find unused or under-used inputs}
\usage{
find_unused_inputs(events, threshold = 0.05)
}
\arguments{
\item{events}{Telemetry events data frame}

\item{threshold}{Percentage threshold for considering input unused}
}
\value{
List of unused input information
}
\description{
Find unused or under-used inputs
}
\keyword{internal}
