% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_suggest.R
\name{get_cognitive_load_suggestions}
\alias{get_cognitive_load_suggestions}
\title{Generate Cognitive Load Theory suggestions}
\usage{
get_cognitive_load_suggestions(previous_stage)
}
\description{
Generate Cognitive Load Theory suggestions
}
\keyword{internal}
