% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{get_concept_bias_mappings}
\alias{get_concept_bias_mappings}
\title{Get bias mitigation strategies for concepts}
\usage{
get_concept_bias_mappings(concepts, mappings = NULL)
}
\arguments{
\item{concepts}{Character vector of concept names}

\item{mappings}{Optional custom concept-bias mappings}
}
\value{
Data frame with relevant bias mappings
}
\description{
Get bias mitigation strategies for concepts
}
