% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_concepts.R
\name{get_concepts_data}
\alias{get_concepts_data}
\title{Internal function to get concepts data from external files}
\usage{
get_concepts_data()
}
\value{
A tibble with all BID framework concepts
}
\description{
Internal function to get concepts data from external files
}
\keyword{internal}
