% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_suggest.R
\name{get_dual_processing_suggestions}
\alias{get_dual_processing_suggestions}
\title{Generate Dual-Processing Theory suggestions}
\usage{
get_dual_processing_suggestions(previous_stage)
}
\description{
Generate Dual-Processing Theory suggestions
}
\keyword{internal}
