% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_suggest.R
\name{get_information_scent_suggestions}
\alias{get_information_scent_suggestions}
\title{Generate Information Scent suggestions}
\usage{
get_information_scent_suggestions(previous_stage)
}
\description{
Generate Information Scent suggestions
}
\keyword{internal}
