% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bid_mappings.R
\name{get_layout_concepts}
\alias{get_layout_concepts}
\title{Get concepts recommended for a layout}
\usage{
get_layout_concepts(layout, mappings = NULL)
}
\arguments{
\item{layout}{Character string indicating layout type}

\item{mappings}{Optional custom layout mappings}
}
\value{
Character vector of recommended concepts
}
\description{
Get concepts recommended for a layout
}
