% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bid-classes.R
\name{get_metadata}
\alias{get_metadata}
\title{Get metadata from bid_stage object}
\usage{
get_metadata(x)
}
\arguments{
\item{x}{A bid_stage object}
}
\value{
List with metadata
}
\description{
Get metadata from bid_stage object
}
