% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_suggest.R
\name{get_progressive_disclosure_suggestions}
\alias{get_progressive_disclosure_suggestions}
\title{Generate Progressive Disclosure suggestions}
\usage{
get_progressive_disclosure_suggestions(previous_stage)
}
\description{
Generate Progressive Disclosure suggestions
}
\keyword{internal}
