% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_suggest.R
\name{get_proximity_suggestions}
\alias{get_proximity_suggestions}
\title{Generate Principle of Proximity suggestions}
\usage{
get_proximity_suggestions(previous_stage)
}
\description{
Generate Principle of Proximity suggestions
}
\keyword{internal}
