% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_classes.R
\name{get_stage}
\alias{get_stage}
\title{Get stage name from bid_stage object}
\usage{
get_stage(x)
}
\arguments{
\item{x}{A bid_stage object}
}
\value{
Character string with stage name
}
\description{
Get stage name from bid_stage object
}
