% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_suggest.R
\name{get_visual_hierarchy_suggestions}
\alias{get_visual_hierarchy_suggestions}
\title{Generate Visual Hierarchy suggestions}
\usage{
get_visual_hierarchy_suggestions(previous_stage)
}
\description{
Generate Visual Hierarchy suggestions
}
\keyword{internal}
