% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_suggest.R
\name{infer_concepts_from_story}
\alias{infer_concepts_from_story}
\title{Infer concepts from Stage 2 (Interpret) story elements}
\usage{
infer_concepts_from_story(previous_stage)
}
\arguments{
\item{previous_stage}{Previous stage data}
}
\value{
Character vector of story-inferred concepts
}
\description{
Infer concepts from Stage 2 (Interpret) story elements
}
\keyword{internal}
