% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_classes.R
\name{is_bid_stage}
\alias{is_bid_stage}
\title{Check if object is a bid_stage}
\usage{
is_bid_stage(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical indicating if object is bid_stage
}
\description{
Check if object is a bid_stage
}
