% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bid-classes.R
\name{is_complete}
\alias{is_complete}
\title{Check if workflow is complete (has all 5 stages)}
\usage{
is_complete(x)
}
\arguments{
\item{x}{A bid_result object}
}
\value{
Logical indicating if workflow is complete
}
\description{
Check if workflow is complete (has all 5 stages)
}
