% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bid_mappings.R
\name{load_concept_bias_mappings}
\alias{load_concept_bias_mappings}
\title{Load concept-bias mappings}
\usage{
load_concept_bias_mappings(custom_mappings = NULL)
}
\arguments{
\item{custom_mappings}{Optional custom mappings data frame}
}
\value{
Data frame with concept-bias mappings
}
\description{
Load concept-bias mappings
}
\keyword{internal}
