% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bid_mappings.R
\name{load_layout_mappings}
\alias{load_layout_mappings}
\title{Load layout-concept mappings}
\usage{
load_layout_mappings(custom_mappings = NULL)
}
\arguments{
\item{custom_mappings}{Optional custom mappings data frame}
}
\value{
Data frame with layout-concept mappings
}
\description{
Load layout-concept mappings
}
\keyword{internal}
