% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bid-classes.R
\name{new_bid_result}
\alias{new_bid_result}
\title{Create a BID result collection object (internal constructor)}
\usage{
new_bid_result(stages)
}
\arguments{
\item{stages}{List of bid_stage objects}
}
\value{
Object of class 'bid_result'
}
\description{
Create a BID result collection object (internal constructor)
}
\keyword{internal}
