% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_classes.R
\name{new_bid_stage}
\alias{new_bid_stage}
\title{Create a BID stage result object (internal constructor)}
\usage{
new_bid_stage(stage, data, metadata = list())
}
\arguments{
\item{stage}{Character string indicating the stage name}

\item{data}{Tibble containing the stage data}

\item{metadata}{List containing additional metadata}
}
\value{
Object of class 'bid_stage'
}
\description{
Create a BID stage result object (internal constructor)
}
\keyword{internal}
