% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_classes.R
\name{print.bid_data_story}
\alias{print.bid_data_story}
\title{Print method for data story objects}
\usage{
\method{print}{bid_data_story}(x, ...)
}
\arguments{
\item{x}{A bid_data_story object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for data story objects
}
