% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bid_classes.R
\name{print.bid_result}
\alias{print.bid_result}
\title{Print method for BID result objects}
\usage{
\method{print}{bid_result}(x, ...)
}
\arguments{
\item{x}{A bid_result object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for BID result objects
}
