% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bid_classes.R
\name{print.bid_stage}
\alias{print.bid_stage}
\title{Print method for BID stage objects}
\usage{
\method{print}{bid_stage}(x, ...)
}
\arguments{
\item{x}{A bid_stage object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for BID stage objects
}
