% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_classes.R
\name{print.bid_user_personas}
\alias{print.bid_user_personas}
\title{Print method for user personas objects}
\usage{
\method{print}{bid_user_personas}(x, ...)
}
\arguments{
\item{x}{A bid_user_personas object}

\item{...}{Additional arguments passed to print.data.frame}
}
\description{
Print method for user personas objects
}
