% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_classes.R
\name{print.bidux_bias_mitigations}
\alias{print.bidux_bias_mitigations}
\title{Print method for bias mitigations objects}
\usage{
\method{print}{bidux_bias_mitigations}(x, ...)
}
\arguments{
\item{x}{A bidux_bias_mitigations object}

\item{...}{Additional arguments passed to print.data.frame}
}
\description{
Print method for bias mitigations objects
}
