% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_suggest.R
\name{rank_and_sort_suggestions}
\alias{rank_and_sort_suggestions}
\title{Rank and sort suggestions within each group}
\usage{
rank_and_sort_suggestions(groups, previous_stage)
}
\arguments{
\item{groups}{List of concept groups with suggestions}

\item{previous_stage}{Previous stage data for scoring adjustments}
}
\value{
List of groups with ranked suggestions
}
\description{
Rank and sort suggestions within each group
}
\keyword{internal}
