% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telemetry_ingest.R
\name{read_telemetry_json}
\alias{read_telemetry_json}
\title{Read telemetry from JSON log file}
\usage{
read_telemetry_json(path)
}
\arguments{
\item{path}{JSON log file path}
}
\value{
Data frame of events
}
\description{
Read telemetry from JSON log file
}
\keyword{internal}
