% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bid_ingest_telemetry.R
\name{read_telemetry_sqlite}
\alias{read_telemetry_sqlite}
\title{Read telemetry from SQLite database}
\usage{
read_telemetry_sqlite(path)
}
\arguments{
\item{path}{SQLite database path}
}
\value{
Data frame of events
}
\description{
Read telemetry from SQLite database
}
\keyword{internal}
