% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_suggest.R
\name{safe_data_story_access}
\alias{safe_data_story_access}
\title{Safe access to data_story elements}
\usage{
safe_data_story_access(data_story, element)
}
\arguments{
\item{element}{Name of data_story element to access}

\item{previous_stage}{Previous stage data}
}
\value{
Character string or empty string if not found
}
\description{
Safe access to data_story elements
}
\keyword{internal}
