% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bid_classes.R
\name{summary.bid_result}
\alias{summary.bid_result}
\title{Summary method for BID result objects}
\usage{
\method{summary}{bid_result}(object, ...)
}
\arguments{
\item{object}{A bid_result object}

\item{...}{Additional arguments (unused)}
}
\description{
Summary method for BID result objects
}
