% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bid-classes.R
\name{validate_bid_result}
\alias{validate_bid_result}
\title{Validate BID result object}
\usage{
validate_bid_result(x)
}
\arguments{
\item{x}{Object to validate}
}
\value{
TRUE if valid, throws error otherwise
}
\description{
Validate BID result object
}
\keyword{internal}
