% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_classes.R
\name{validate_bid_stage}
\alias{validate_bid_stage}
\title{Validate BID stage object}
\usage{
validate_bid_stage(x)
}
\arguments{
\item{x}{Object to validate}
}
\value{
TRUE if valid, throws error otherwise
}
\description{
Validate BID stage object
}
\keyword{internal}
