% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bid_concepts.R
\name{bid_concepts}
\alias{bid_concepts}
\title{List all BID Framework Concepts}
\usage{
bid_concepts()
}
\value{
A tibble with concept details.
}
\description{
This function returns a tibble listing all the key concepts in the Behavior Insight Design framework,
including their descriptions, categories, references, and example applications.
}
\examples{
bid_concepts()
}
