% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bid_with_quiet}
\alias{bid_with_quiet}
\title{Temporarily suppress bidux messages}
\usage{
bid_with_quiet(code)
}
\arguments{
\item{code}{Code to execute with messages suppressed}
}
\value{
The result of evaluating code
}
\description{
Execute code with bidux messages temporarily suppressed.
}
\examples{
# Run analysis quietly without changing global setting
result <- bid_with_quiet({
  bid_interpret(
    central_question = "How can we improve user engagement?",
    data_story = list(hook = "Users are leaving", resolution = "Fix issues")
  )
})

}
