% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bid_package.R
\docType{package}
\name{bidux-package}
\alias{bidux}
\alias{bidux-package}
\title{bidux: Behavioral Insight Design: A Toolkit for Integrating Behavioral Science in UI/UX Design}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a framework and toolkit to guide R dashboard developers in implementing the Behavioral Insight Design (BID) framework. The package offers functions for documenting each of the five stages (Interpret, Notice, Anticipate, Structure, and Validate), along with a comprehensive concept dictionary. Works with both 'shiny' applications and 'Quarto' dashboards.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jrwinget.github.io/bidux/}
}

}
\author{
\strong{Maintainer}: Jeremy Winget \email{contact@jrwinget.com} (\href{https://orcid.org/0000-0002-3783-4354}{ORCID})

}
\keyword{internal}
