% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_suggest.R
\name{build_groups_with_suggestions}
\alias{build_groups_with_suggestions}
\title{Build suggestion groups organized by concept}
\usage{
build_groups_with_suggestions(concepts_final, chosen_layout, previous_stage)
}
\arguments{
\item{concepts_final}{Final list of concepts to generate suggestions for}

\item{chosen_layout}{Selected layout type}

\item{previous_stage}{Previous stage data for context}
}
\value{
List of concept groups with suggestions
}
\description{
Build suggestion groups organized by concept
}
\keyword{internal}
