% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telemetry.R
\name{create_confusion_notice}
\alias{create_confusion_notice}
\title{Create notice stage for confusion patterns}
\usage{
create_confusion_notice(confusion_info, total_sessions)
}
\arguments{
\item{confusion_info}{List with confusion pattern information}

\item{total_sessions}{Total number of sessions}
}
\value{
bid_stage object
}
\description{
Create notice stage for confusion patterns
}
\keyword{internal}
