% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telemetry_notice_creators.R
\name{create_unused_input_notice}
\alias{create_unused_input_notice}
\title{Create notice stage for unused input}
\usage{
create_unused_input_notice(input_info, total_sessions)
}
\arguments{
\item{input_info}{List with input usage information}

\item{total_sessions}{Total number of sessions}
}
\value{
bid_stage object
}
\description{
Create notice stage for unused input
}
\keyword{internal}
