% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_suggest.R
\name{get_generic_suggestions}
\alias{get_generic_suggestions}
\title{Generate generic suggestions for unrecognized concepts}
\usage{
get_generic_suggestions(concept, chosen_layout, previous_stage)
}
\description{
Generate generic suggestions for unrecognized concepts
}
\keyword{internal}
