% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_rank.R
\name{get_onboarding_suggestions}
\alias{get_onboarding_suggestions}
\title{Generate User Onboarding suggestions}
\usage{
get_onboarding_suggestions(chosen_layout, previous_stage)
}
\description{
Generate User Onboarding suggestions
}
\keyword{internal}
