% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telemetry.R
\name{read_telemetry_data}
\alias{read_telemetry_data}
\title{Read telemetry data from file}
\usage{
read_telemetry_data(path, format)
}
\arguments{
\item{path}{File path}

\item{format}{Format ("sqlite" or "json")}
}
\value{
Data frame of events
}
\description{
Read telemetry data from file
}
\keyword{internal}
