% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telemetry_ingest.R
\name{read_telemetry_data}
\alias{read_telemetry_data}
\title{Read telemetry data from file}
\usage{
read_telemetry_data(path, format, events_table = NULL)
}
\arguments{
\item{path}{File path}

\item{format}{Format ("sqlite" or "json")}

\item{events_table}{Optional custom events table for SQLite}
}
\value{
Data frame of events
}
\description{
Read telemetry data from file
}
\keyword{internal}
