% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telemetry.R
\name{read_telemetry_sqlite}
\alias{read_telemetry_sqlite}
\title{Read telemetry from SQLite database}
\usage{
read_telemetry_sqlite(path, events_table = NULL)
}
\arguments{
\item{path}{SQLite database path}

\item{events_table}{Optional custom events table data.frame}
}
\value{
Data frame of events
}
\description{
Read telemetry from SQLite database
}
\keyword{internal}
